requirements.txt
streamlit>=1.32
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-classic>=1.0.0
langchain-chroma>=1.1.0
chromadb>=1.4.0
pypdf>=6.0.0
